requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
requests>=2.32.3
Markdown>=3.7
PyYAML>=6.0.2
python-multipart>=0.0.12
chromadb>=0.5.20
sentence-transformers>=3.3.1
PyGithub>=2.5.0
pytest>=8.3.4
httpx>=0.28.1
pydantic-settings>=2.6.1
aiohttp>=3.11.11